refactor: remove redudant code in yield claim flow
