+++ packages/suite/src/components/earn/dashboard/yield/EarnYieldClaimSelectAccountModal.tsx
@@ -1,7 +1,6 @@
 import { events } from '@suite/analytics';
 import { Translation } from '@suite/intl';
 import { useFormatters } from '@suite-common/formatters';
-import { type Account, type BaseCurrencyAmount } from '@suite-common/wallet-types';
 import { CardList, Column, Modal, Row, Text } from '@trezor/components';
 import { CoinLogo } from '@trezor/product-components';
 
@@ -9,37 +8,34 @@ import { AccountLabel } from 'src/components/suite/AccountLabel';
 import { Address } from 'src/components/suite/Address';
 import { useAnalytics } from 'src/support/useAnalytics';
 
-export type EarnYieldClaimableAccount = {
-    account: Account;
-    totalFiatAmount: BaseCurrencyAmount | null;
-};
+import { type YieldAccountRewards, type YieldAccountsRewards } from '../../yield/claim/hooks';
 
 type EarnYieldClaimSelectAccountModalProps = {
-    claimableAccounts: EarnYieldClaimableAccount[];
-    onSelect: (claimableAccount: EarnYieldClaimableAccount) => void;
+    accountsRewards: YieldAccountsRewards;
+    onSelect: (account: YieldAccountRewards) => void;
     onClose: () => void;
 };
 
 export const EarnYieldClaimSelectAccountModal = ({
-    claimableAccounts,
+    accountsRewards,
     onSelect,
     onClose,
 }: EarnYieldClaimSelectAccountModalProps) => {
     const analytics = useAnalytics();
     const { BaseCurrencyAmountFormatter } = useFormatters();
 
-    const handleOnSelect = (claimableAccount: EarnYieldClaimableAccount) => {
+    const handleOnSelect = (account: YieldAccountRewards) => {
         analytics.report({
             type: events.yieldNavigateEvent.name,
             payload: {
                 action: 'continue',
                 from: 'claim-select-account-modal',
                 to: 'claim-form',
-                networkSymbol: claimableAccount.account.symbol,
+                networkSymbol: account.account.symbol,
             },
         });
 
-        onSelect(claimableAccount);
+        onSelect(account);
     };
 
     const handleOnCancel = () => {
@@ -62,20 +58,24 @@ export const EarnYieldClaimSelectAccountModal = ({
             onCancel={handleOnCancel}
         >
             <CardList>
-                {claimableAccounts.map(claimable => (
+                {accountsRewards.map(accountRewards => (
                     <CardList.Item
-                        key={claimable.account.key}
-                        onClick={() => handleOnSelect(claimable)}
+                        key={accountRewards.account.key}
+                        onClick={() => handleOnSelect(accountRewards)}
                     >
                         <Row gap={16} flex="1" overflow="hidden">
-                            <CoinLogo symbol={claimable.account.symbol} size={32} type="token" />
+                            <CoinLogo
+                                symbol={accountRewards.account.symbol}
+                                size={32}
+                                type="token"
+                            />
                             <Column flex="1" overflow="hidden" gap={2} alignItems="flex-start">
                                 <AccountLabel
-                                    account={claimable.account}
+                                    account={accountRewards.account}
                                     typographyStyle="body-md-strong"
                                 />
                                 <Address
-                                    value={claimable.account.descriptor}
+                                    value={accountRewards.account.descriptor}
                                     typographyStyle="body-sm"
                                     intent="neutral"
                                     priority="secondary"
@@ -84,9 +84,9 @@ export const EarnYieldClaimSelectAccountModal = ({
                             </Column>
                         </Row>
                         <Text typographyStyle="body-md-strong">
-                            {claimable.totalFiatAmount !== null
-                                ? BaseCurrencyAmountFormatter.format(claimable.totalFiatAmount)
-                                : '—'}
+                            {BaseCurrencyAmountFormatter.format(
+                                accountRewards.totalClaimableFiatAmount,
+                            )}
                         </Text>
                     </CardList.Item>
                 ))}
