+++ packages/suite/src/components/earn/dashboard/yield/hooks/useYieldTableData.ts
@@ -209,10 +209,16 @@ export const useYieldTableData = ({
         [yieldAccountOpportunities],
     );
 
+    const yieldAccounts = useMemo(
+        () => yieldAccountOpportunities.map(opportunity => opportunity.account),
+        [yieldAccountOpportunities],
+    );
+
     return {
         isYieldActive,
         hasAnyRewardsData,
         yieldAccountOpportunities,
         yieldInactiveVaultOpportunities,
+        yieldAccounts,
     };
 };
