+++ packages/suite/src/components/earn/yield/claim/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useMerkleRewards';
